requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1707233003401/work
arxiv==2.1.0
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
attrs==23.2.0
autopep8==2.1.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
coloredlogs==15.0.1
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.2.1
cryptography==42.0.8
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.4
debugpy @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/debugpy_1699253073094/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
effdet==0.4.1
email_validator==2.1.1
emoji==2.12.1
et-xmlfile==1.1.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1704921103267/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.8.0
fastapi==0.111.0
fastapi-cli==0.0.4
feedparser==6.0.10
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-vision==3.7.2
google-search-results==2.4.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.6
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1710971335535/work
importlib_resources==6.4.0
iopath==0.1.10
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1717717537056/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1717182742060/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jupyter_core_1701803168398/work
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.8
langchain-experimental==0.0.61
langchain-openai==0.1.13
langchain-postgres==0.0.6
langchain-text-splitters==0.2.1
langchainhub==0.1.15
langdetect==1.0.9
langgraph==0.0.69
langserve==0.2.1
langsmith==0.1.78
layoutparser==0.3.4
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.0
openai==1.34.0
opencv-python==4.10.0.82
openpyxl==3.1.3
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
peewee==3.17.5
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pgvector==0.2.5
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pikepdf==9.0.0
pillow==10.3.0
pillow_heif==0.16.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
playwright==1.44.0
polygon-api-client==1.13.7
portalocker==2.8.2
posthog==3.5.0
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
proto-plus==1.23.0
protobuf==4.25.3
psutil @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/psutil_1699248249804/work
psycopg==3.1.19
psycopg-pool==3.2.2
psycopg2-binary==2.9.9
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.22
pydantic==1.10.13
pydantic_core==2.16.3
pydeck==0.9.1
pyee==11.1.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject-toml==0.0.10
pyproject_hooks==1.1.0
pytesseract==0.3.10
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
pyzmq @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_43pxpbos3z/croot/pyzmq_1705605108344/work
rapidfuzz==3.9.3
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scipy==1.13.1
setuptools==68.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sse-starlette==2.1.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.37.2
streamlit==1.34.0
sympy==1.12.1
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.7.0
timm==1.0.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/tornado_1699243541634/work
tqdm==4.66.2
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
transformers==4.41.2
typer==0.12.3
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.5
unstructured-client==0.23.3
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.12
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.41.2
wikipedia==1.4.0
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.4
yfinance==0.2.38
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1718013267051/work
